requirements.txt
torch
timm
Pillow
opencv-python
numpy
